new file foc classes
